🚚 rename: 티켓 취소 및 예매내역 가져오는 로직을 useMyTicketHistory 커스텀 훅으로 분리 #46
